auth_api/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiofiles==24.1.0
    # via nicegui
aiohappyeyeballs==2.6.1
    # via
    #   auth-api (pyproject.toml)
    #   aiohttp
aiohttp==3.11.16
    # via
    #   auth-api (pyproject.toml)
    #   elasticsearch
    #   nicegui
    #   python-socketio
aiosignal==1.3.2
    # via aiohttp
alembic==1.14.1
    # via auth-api (pyproject.toml)
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   auth-api (pyproject.toml)
    #   httpx
    #   starlette
    #   watchfiles
asgiref==3.8.1
    # via opentelemetry-instrumentation-asgi
async-fastapi-jwt-auth==0.6.6
    # via auth-api (pyproject.toml)
asyncpg==0.30.0
    # via auth-api (pyproject.toml)
attrs==25.3.0
    # via
    #   auth-api (pyproject.toml)
    #   aiohttp
backoff==2.2.1
    # via auth-api (pyproject.toml)
bidict==0.23.1
    # via python-socketio
certifi==2024.12.14
    # via
    #   elastic-transport
    #   httpcore
    #   httpx
    #   nicegui
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.2
    # via requests
click==8.1.8
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
deprecated==1.2.18
    # via
    #   opentelemetry-api
    #   opentelemetry-semantic-conventions
dnspython==2.7.0
    # via email-validator
docutils==0.21.2
    # via nicegui
elastic-transport==8.17.0
    # via elasticsearch
elasticsearch==8.13.2
    # via auth-api (pyproject.toml)
email-validator==2.2.0
    # via
    #   auth-api (pyproject.toml)
    #   fastapi
    #   pydantic
fastapi==0.115.11
    # via
    #   auth-api (pyproject.toml)
    #   async-fastapi-jwt-auth
    #   fastapi-limiter
    #   fastapi-sso
    #   nicegui
fastapi-cli==0.0.7
    # via fastapi
fastapi-limiter==0.1.6
    # via auth-api (pyproject.toml)
fastapi-sso==0.18.0
    # via auth-api (pyproject.toml)
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
googleapis-common-protos==1.59.1
    # via opentelemetry-exporter-jaeger-proto-grpc
greenlet==3.1.1
    # via sqlalchemy
grpcio==1.71.0
    # via opentelemetry-exporter-jaeger-proto-grpc
h11==0.16.0
    # via
    #   httpcore
    #   nicegui
    #   uvicorn
    #   wsproto
httpcore==1.0.9
    # via
    #   auth-api (pyproject.toml)
    #   httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   async-fastapi-jwt-auth
    #   fastapi
    #   fastapi-sso
    #   nicegui
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
ifaddr==0.2.0
    # via nicegui
importlib-metadata==8.6.1
    # via opentelemetry-api
itsdangerous==2.2.0
    # via
    #   auth-api (pyproject.toml)
    #   nicegui
jinja2==3.1.6
    # via
    #   fastapi
    #   nicegui
mako==1.3.9
    # via alembic
markdown-it-py==3.0.0
    # via rich
markdown2==2.5.3
    # via nicegui
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
nicegui==2.18.0
    # via auth-api (pyproject.toml)
oauthlib==3.2.2
    # via fastapi-sso
opentelemetry-api==1.31.0
    # via
    #   auth-api (pyproject.toml)
    #   opentelemetry-exporter-jaeger-proto-grpc
    #   opentelemetry-exporter-jaeger-thrift
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-sqlalchemy
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-jaeger==1.21.0
    # via auth-api (pyproject.toml)
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
    # via opentelemetry-exporter-jaeger
opentelemetry-exporter-jaeger-thrift==1.21.0
    # via opentelemetry-exporter-jaeger
opentelemetry-instrumentation==0.52b0
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-asgi==0.52b0
    # via opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-fastapi==0.52b0
    # via auth-api (pyproject.toml)
opentelemetry-instrumentation-sqlalchemy==0.52b0
    # via auth-api (pyproject.toml)
opentelemetry-sdk==1.31.0
    # via
    #   auth-api (pyproject.toml)
    #   opentelemetry-exporter-jaeger-proto-grpc
    #   opentelemetry-exporter-jaeger-thrift
opentelemetry-semantic-conventions==0.52b0
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-sqlalchemy
    #   opentelemetry-sdk
opentelemetry-util-http==0.52b0
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
orjson==3.10.15
    # via
    #   auth-api (pyproject.toml)
    #   nicegui
packaging==24.2
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-sqlalchemy
passlib==1.7.4
    # via auth-api (pyproject.toml)
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
protobuf==4.25.6
    # via googleapis-common-protos
pscript==0.7.7
    # via vbuild
psycopg==3.2.4
    # via auth-api (pyproject.toml)
psycopg-binary==3.2.4
    # via psycopg
psycopg2-binary==2.9.10
    # via auth-api (pyproject.toml)
pycryptodomex==3.21.0
    # via auth-api (pyproject.toml)
pydantic==2.10.5
    # via
    #   fastapi
    #   fastapi-sso
    #   pydantic-settings
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.7.1
    # via async-fastapi-jwt-auth
pygments==2.19.1
    # via
    #   nicegui
    #   rich
pyjwt==2.10.1
    # via
    #   async-fastapi-jwt-auth
    #   fastapi-sso
python-dateutil==2.9.0
    # via auth-api (pyproject.toml)
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
python-engineio==4.12.1
    # via
    #   nicegui
    #   python-socketio
python-json-logger==3.3.0
    # via auth-api (pyproject.toml)
python-multipart==0.0.20
    # via
    #   auth-api (pyproject.toml)
    #   fastapi
    #   nicegui
python-socketio==5.13.0
    # via nicegui
pyyaml==6.0.2
    # via uvicorn
redis==5.2.1
    # via
    #   auth-api (pyproject.toml)
    #   fastapi-limiter
requests==2.32.3
    # via nicegui
rich==13.9.4
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.13.2
    # via fastapi-cli
sentry-sdk==1.39.1
    # via auth-api (pyproject.toml)
shellingham==1.5.4
    # via typer
simple-websocket==1.1.0
    # via python-engineio
six==1.17.0
    # via
    #   python-dateutil
    #   thrift
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.38
    # via
    #   auth-api (pyproject.toml)
    #   alembic
starlette==0.46.0
    # via
    #   fastapi
    #   nicegui
thrift==0.21.0
    # via opentelemetry-exporter-jaeger-thrift
typer==0.15.1
    # via
    #   auth-api (pyproject.toml)
    #   fastapi-cli
typing-extensions==4.12.2
    # via
    #   alembic
    #   fastapi
    #   nicegui
    #   opentelemetry-sdk
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   sqlalchemy
    #   typer
urllib3==2.3.0
    # via
    #   elastic-transport
    #   nicegui
    #   requests
    #   sentry-sdk
uvicorn==0.34.0
    # via
    #   fastapi
    #   fastapi-cli
    #   nicegui
uvloop==0.21.0
    # via uvicorn
vbuild==0.8.2
    # via nicegui
wait-for2==0.3.2
    # via nicegui
watchfiles==1.0.4
    # via
    #   nicegui
    #   uvicorn
websockets==15.0
    # via uvicorn
werkzeug==3.1.3
    # via auth-api (pyproject.toml)
wrapt==1.17.2
    # via
    #   deprecated
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-sqlalchemy
wsproto==1.2.0
    # via simple-websocket
yarl==1.18.3
    # via aiohttp
zipp==3.21.0
    # via importlib-metadata
